+++ cmd/zetae2e/local/local.go
@@ -3,12 +3,10 @@ package local
 import (
 	"context"
 	"errors"
-	"fmt"
 	"os"
 	"path/filepath"
 	"time"
 
-	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/fatih/color"
 	"github.com/spf13/cobra"
 	"github.com/stretchr/testify/require"
@@ -353,7 +351,7 @@ func localE2ETest(cmd *cobra.Command, _ []string) {
 	logger.Print("✅ e2e tests completed in %s", time.Since(testStartTime).String())
 
 	if testTSSMigration {
-		runTSSMigrationTest(deployerRunner, logger, verbose, conf)
+		runTSSMigration(deployerRunner, logger, verbose, conf)
 	}
 	// Verify that there are no trackers left over after tests complete
 	deployerRunner.EnsureNoTrackers()
@@ -408,57 +406,7 @@ func waitKeygenHeight(
 	}
 }
 
-func updateTssAddressForConnector(runner *runner.E2ERunner) {
-
-	tssAddress, err := runner.ConnectorEth.TssAddress(&bind.CallOpts{Context: runner.Ctx})
-	require.NoError(runner, err)
-	runner.Logger.Print(fmt.Sprintf("TSS Address Before: %s", tssAddress.String()))
-
-	tssUpdater, err := runner.ConnectorEth.TssAddressUpdater(&bind.CallOpts{Context: runner.Ctx})
-	require.NoError(runner, err)
-	runner.Logger.Print(fmt.Sprintf("TSS Updater: %s", tssUpdater.String()))
-
-	runner.Logger.Print("Update TSS")
-	noError(runner.SetTSSAddresses())
-	runner.Logger.Print("TSS Deployer Address: %s", runner.TSSAddress)
-
-	tx, err := runner.ConnectorEth.UpdateTssAddress(runner.EVMAuth, runner.TSSAddress)
-	require.NoError(runner, err)
-	runner.Logger.Print(fmt.Sprintf("TSS Address Update Tx: %s", tx.Hash().String()))
-	receipt := utils.MustWaitForTxReceipt(runner.Ctx, runner.EVMClient, tx, runner.Logger, runner.ReceiptTimeout)
-	utils.RequireTxSuccessful(runner, receipt)
-
-	tssAddress, err = runner.ConnectorEth.TssAddress(&bind.CallOpts{Context: runner.Ctx})
-	require.NoError(runner, err)
-	runner.Logger.Print(fmt.Sprintf("TSS Address After: %s", tssAddress.String()))
-}
-
-func updateTssAddressForErc20custody(runner *runner.E2ERunner) {
-
-	tssAddress, err := runner.ERC20Custody.TSSAddress(&bind.CallOpts{Context: runner.Ctx})
-	require.NoError(runner, err)
-	runner.Logger.Print(fmt.Sprintf("TSS ERC20 Address Before: %s", tssAddress.String()))
-
-	tssUpdater, err := runner.ERC20Custody.TSSAddressUpdater(&bind.CallOpts{Context: runner.Ctx})
-	require.NoError(runner, err)
-	runner.Logger.Print(fmt.Sprintf("TSS ERC20 Updater: %s", tssUpdater.String()))
-
-	runner.Logger.Print("Update TSS")
-	noError(runner.SetTSSAddresses())
-	runner.Logger.Print("TSS Deployer Address: %s", runner.TSSAddress)
-
-	tx, err := runner.ERC20Custody.UpdateTSSAddress(runner.EVMAuth, runner.TSSAddress)
-	require.NoError(runner, err)
-	runner.Logger.Print(fmt.Sprintf("TSS ERC20 Address Update Tx: %s", tx.Hash().String()))
-	receipt := utils.MustWaitForTxReceipt(runner.Ctx, runner.EVMClient, tx, runner.Logger, runner.ReceiptTimeout)
-	utils.RequireTxSuccessful(runner, receipt)
-
-	tssAddress, err = runner.ERC20Custody.TSSAddress(&bind.CallOpts{Context: runner.Ctx})
-	require.NoError(runner, err)
-	runner.Logger.Print(fmt.Sprintf("TSS ERC20 Address After: %s", tssAddress.String()))
-}
-
-func runTSSMigrationTest(deployerRunner *runner.E2ERunner, logger *runner.Logger, verbose bool, conf config.Config) {
+func runTSSMigration(deployerRunner *runner.E2ERunner, logger *runner.Logger, verbose bool, conf config.Config) {
 	migrationStartTime := time.Now()
 	logger.Print("🏁 starting tss migration")
 
@@ -483,26 +431,9 @@ func runTSSMigrationTest(deployerRunner *runner.E2ERunner, logger *runner.Logger
 		logger.Print("❌ tss migration failed")
 		os.Exit(1)
 	}
-	updateTssAddressForConnector(deployerRunner)
-	updateTssAddressForErc20custody(deployerRunner)
+	deployerRunner.UpdateTssAddressForConnector()
+	deployerRunner.UpdateTssAddressForErc20custody()
 	logger.Print("✅ migration completed in %s ", time.Since(migrationStartTime).String())
-
-	//logger.Print("🏁 starting post migration tests")
-
-	//tests := []string{
-	//	e2etests.TestBitcoinWithdrawSegWitName,
-	//	e2etests.TestEtherWithdrawName,
-	//	e2etests.TestZetaWithdrawName,
-	//	e2etests.TestERC20WithdrawName,
-	//}
-	//fn = postMigrationTestRoutine(conf, deployerRunner, verbose, tests...)
-	//
-	//if err := fn(); err != nil {
-	//	logger.Print("❌ %v", err)
-	//	logger.Print("❌ post migration tests failed")
-	//	os.Exit(1)
-	//}
-
 }
 
 func must[T any](v T, err error) T {