+++ x/crosschain/types/expected_keepers.go
@@ -103,8 +103,6 @@ type ObserverKeeper interface {
 	) (bool, bool, observertypes.Ballot, string, error)
 	GetSupportedChainFromChainID(ctx sdk.Context, chainID int64) (chains.Chain, bool)
 	GetSupportedChains(ctx sdk.Context) []chains.Chain
-	GetSupportedForeignChains(ctx sdk.Context) []chains.Chain
-	GetSupportedForeignChainsByConsensus(ctx sdk.Context, consensus chains.Consensus) []chains.Chain
 }
 
 type FungibleKeeper interface {